fully ended Time sheet project, but code not simplified yet

+++ src/main/java/timesheet/TimeSheetItem.java
@@ -33,7 +33,7 @@ public class TimeSheetItem {
     }
 
     public long hoursBetweenDates() {
-        return Duration.between(beginDate, endDate).toHours();
+        return (long) Duration.between(beginDate, endDate).toHours();
     }
 
     public Employee getEmployee() {
